perspective correct Z new rendering algorithm new lerp some color utils some vector utils

+++ PointsRenderer.hpp
@@ -8,5 +8,5 @@
 
 namespace PointsRenderer
 {
-	void renderPoints(const std::vector<Point>& points, Drawable* drawSurface, Matrix2D<int>& zBuffer, const Rect& viewPort, const Camera& camera);
+	void renderPoints(const std::vector<Point>& points, Drawable* drawSurface, Matrix2D<double>& zBuffer, const Rect& viewPort, const Camera& camera);
 }
